renaming folder albums for netlify compatibility
