+++ src/components/albums/album-child-component/CoverImage.jsx
